Remove almost empty files, move commands, config in tx_run
