+++ rust-client/src/lib.rs
@@ -1,4 +1,2 @@
-pub mod command;
-pub mod config;
 pub mod gas;
 pub mod tx_run;
